todoapp_backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
backports.zoneinfo==0.2.1
django==4.2.25
django-ninja==1.4.3
mssql-django==1.6
pydantic==2.10.6
pydantic-core==2.27.2
pyodbc==5.2.0
pytz==2025.2
sqlparse==0.5.3
typing-extensions==4.13.2